new change from ter,dorm station, added only time phaser to shared :>

+++ mandatory/main/utils.h
@@ -15,7 +15,7 @@
 
 #include <limits.h>
 
-long				ft_atoi(char *str);
+size_t				ft_atoi(char *str);
 void                error_message(char *text, int signal);
 int                 ft_strlen(char *str);
